chore(allure): fixing shared config file to dont overlap
