+++ config/wdio.windows.multiremote.conf.ts
@@ -1,6 +1,5 @@
 import allureReporter from '@wdio/allure-reporter'
 import { config as sharedConfig } from './wdio.shared.conf';
-import { homedir } from "os";
 import { join } from "path";
 const fsp = require("fs").promises;
 const userACacheFolder = join(process.cwd(), "./apps/ChatUserA/.user")
@@ -87,6 +86,18 @@ export const config: WebdriverIO.Config = {
     // The only one supported by default is 'dot'
     // see also: https://webdriver.io/docs/dot-reporter
     reporters: [
+      ["spec", 
+        {
+          showPreface: false,
+        },
+      ], 
+      ['allure', 
+        {
+          outputDir: './allure-results',
+          disableWebdriverStepsReporting: true,
+          disableWebdriverScreenshotsReporting: false,
+        }
+      ],
       ['junit', 
         {
           outputDir: './test-report/',
@@ -107,6 +118,19 @@ export const config: WebdriverIO.Config = {
     onPrepare: async function() {
       const cacheFolderUserA = join(process.cwd(), "\\apps\\ChatUserB")
       const cacheFolderUserB = join(process.cwd(), "\\apps\\ChatUserB")
+      const allureResultsFolder = join(process.cwd(), "\\allure-results");
+      const testReportFolder =  join(process.cwd(), "\\test-report");
+      const testResultsFolder =  join(process.cwd(), "\\test-results");
+      try {
+        await rmSync(allureResultsFolder, { recursive: true, force: true });
+        await rmSync(testReportFolder, { recursive: true, force: true });
+        await rmSync(testResultsFolder, { recursive: true, force: true });
+        console.log("Deleted Artifacts Folders Successfully!");
+      } catch (error) {
+        console.error(
+            `Got an error trying to delete artifacts folders: ${error.message}`
+        );
+      }
       try {
         await rmSync(cacheFolderUserA, { recursive: true, force: true });
         await rmSync(cacheFolderUserB, { recursive: true, force: true });
